Don't return a 400 status code on errors
The GraphQL standard does not define the HTTP status code that should be
returned on errors, but the libraries tend to return 200 and use the
`errors` dictionary to signal errors.

Apollo will consider that 4xx and 5xx codes are errors and will not
process the result as it would, reporting errors in the `errors` dict.

I can think of two other good reasons to return a 200 code on errors:

- GraphQL allows partial results, a response is thus not only good or
  bad
- Even if it's not supported right now by this library, GraphQL allows
  multiple queries in the same request. If one queries succeeds and the
  other fails, it seems incorrect to return a 400 error code.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>